LayoutTests/accessibility/html-html-element-is-ignored-expected.txt

This test checks that the element is not the AX tree (ever), even when it has an aria attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The start element should be a button.
PASS startElement.role is 'AXRole: AXButton'

The parent should be the web area.
PASS startElement.parentElement().role is 'AXRole: AXWebArea'
PASS successfullyParsed is true

TEST COMPLETE
